+++ src/frontend/apps/impress/src/features/docs/members/members-add/components/index.ts
@@ -1,1 +1,2 @@
 export * from './AddMembers';
+export * from './ChooseRole';